requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==20.8b1
boto3==1.17.47
botocore==1.20.47
certifi==2020.12.5
chardet==4.0.0
chevron==0.14.0
click==7.1.2
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
gunicorn==20.1.0
idna==2.10
image==1.5.33
jmespath==0.10.0
lxml==4.6.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
PyJWT==2.0.1
python-dateutil==2.1
python-decouple==3.4
pytz==2021.1
regex==2021.4.4
requests==2.25.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
whitenoise==5.2.0
